[SIMB-1830] easycwmp: download size check before firmware update
Ensure we abort any firmware update attempts when we detect that the received
object is not exactly the same size we expected to receive.

While at it, optimize and modernize this area of the shell code, and limit the
download size to 90% of /tmp, instead of 100%: we don't want to starve the box
before applying the firmware image (or worse, *during* the sysupgrade process).

Upstream-report: http://support.easycwmp.org/view.php?id=138