[SIMB-1898] simetbox-openwrt-base: sysupgrade glue for 32MiB RAM
Add in simetbox-openwrt-feed (because that's where the openwrt glue
should be) /lib/upgrade scripts to make sysupgrade safer on low-ram
routers, such as the TP-Link TL-WR842NDv2.

It tries harder to free RAM, and stops the update (with a reboot before
any flash writing is done) if there are less than 10000 KiB free for the
do_upgrade() procedure.