[SIMB-1898] adjust kernel's OOM killer score
Adjust the OOM killer score so that it will try to kill everything
else but sysupgrade and its children.

This change is supposed to be coupled to a change in other simetbox
packages to make it more likely for measurements and other less
important components to be killed first.

OpenWRT processes typically have an offset of zero.  We use -500 at the
start of the sysupgrade process, and -1000 (disable the OOM killer
entirely for sysupgrade and its children) when we are about to call
mtd write.