[SIMB-1898] simetbox-openwrt-base: lowmem sysupgrade: harden killall
killall may return a non-zero exit status if it did not manage to kill
every process (1) or if it killed no process at all (2).  Ensure we
are not going to abort with an error in that case, even if "set -e" is
in effect

Note that, from testing, it appears that "set -e" is _not_ in effect,
but this is a just-in-case change.