[MDC-165] sysupgrade: TP-Link: blink power/status + qss/wps LEDS
Powering down the device during firmware flash would cause a misflash
likely requiring bootloader-based firmware recovery (TFTP, etc), and if
one is very unlucky, hardware damage to the FLASH.

On TP-Link devices, blink the power/status and the qss/wps LEDs when we
are about to start flashing the device on sysupgrade.  We don't blink
every LED we can because that would look like a faulty unit.

Hopefully, this will be enough to hint to the user to not power down the
device until we finish flashing and reboot (which takes less than a
minute).