[MDC-434] simetbox-openwrt-feed: add a package for simet-lmapd
Add the new lmapd scheduler as simetbox-openwrt-simet-lmapd.  Build
lmapd with shared library support and without XML support (MDC-438), to
reduce FLASH requirements (libxml2 is very large, we don't want it as a
dependency).

All the configuration goes into simetbox-openwrt-simet-ma, the new
package is there just because of the separate upstream nature of
simet-lmapd.

Add a procd service (and start it on boot) for simet-lmapd, run as the
unprivileged nicbr-simet user for now (we *may* have to use root later)
and in JSON mode.

Use multiple config paths (hardcoded at compile time, overrideable at
runtime).  We need this for simet-ma to be able to easily validate a
config change candidate (MDC-452).