[MDC-637] simetbox-openwrt-simet-ma: ensure sha256sum is available
Unfortunately, simet-ma 0.14.3 and 0.14.4 require sha256sum for the
simetbox_fetch-lmap-schedule script.

Ensure some version of sha256sum will be available.

This has been fixed on later versions of simet-ma to use md5sum instead,
which is available by default on older versions of OpenWRT, and the
DEPENDS should be changed to match.