[MDC-650] simetbox-openwrt-core: remove package
Unfortunately, the "bcp38" package from OpenWRT is "misguided" and
"misnamed" to put it nicely.  It has nothing to do with spoofed source
address filtering, anyway.

And it definately does *not* implement BCP-38 requirements and
recommendations.  We need to replace it with something that does.

Since trying to enable OpenWRT's bcp38 package is the only thing that
the simetbox-openwrt-core was doing ATM, remove the package entirely.
We can reintroduce it later if needed.

Note that NIC.br has never shipped firmware with the
simetbox-openwrt-core package enabled.