[MDC-829] simetbox-openwrt-luci: be clear about background work
Instead of reloading the page, leave the notice about the device doing
something in background *if* an update is available.

This is a bit more clear about how that button works, but the real fix
for this usability issue is to actually have the background process
signal something to the UI and have that show on the status line,
regardless of whether the update check was requested by cron or button.