[MDC-826] simetbox-openwrt-luci: add entrypoints for measurement runs
Add LuCI entrypoints to access "ubus simet_ma simet_engine_status",
and "ubus simet_ma start_measurement_run".

This allows us to retrieve wether the SIMET2 engine is active (right now
it always is), and whether it is currently doing a measurement run, as
well as request it to start a measurement run (if one is not already
running).