[MDC-859] simetbox-openwrt-simet-ma: really randomize crontab times
AWK srand() uses the current system time (and apparently at a low
granularity, like 1s), unless you give it a seed (apparently, a 31-bit
positive integer).

uci-defaults scripts run very early, and on the usual WiSoC without a
RTC, the system time at the moment these scripts get to run is *quite*
deterministic on the same model with the same firmware.

Because of that, AWK srand() was generating the *same* seed on a very
large fraction of the routers, and all the crontab start times (adjusted
by an uci-defaults script), which we wanted to spread to avoid
clustering, ended up group-sync'd and highly clustered.

Fix it trying to use the lower 5 digits of the MAC address as the seed,
this has enough variation (further spread by srand()) to mostly remove
the clustering.