[MDC-770] simetbox-openwrt-luci: require sysauth always
We can hit a very nasty OpenWRT-19.07.x+ issue with LuCI should we set
sysauth=false on a *page* (not a XHR node, etc).  Basically, the menu
diappears, and you cannot navigate anywhere.  Refer to
https://github.com/openwrt/luci/issues/3563

It was somewhat fixed on LuCI master (for OpenWRT 20), but not in a way
that made people happy -- since it will always show the entire menu
tree regardless of auth state.  But OpenWRT 19.07 doesn't have this
change, it is too controrversial to request a backport, and we don't
want to fork LuCI if we can help it.

While it is really nice to have a decent status page that does not
require auth (something OpenWRT sorely lacks is user friendliness), we
will have to address that directly later.

While at it, remove other stray code that is currently unused (and has
been unused for a looong time).

Note that we do not have the "user_data" functionality active for
*years*, but remants of its code are still everywhere...