[MDC-882] simetbox-openwrt-simet-ma: support inetup netdev monitoring
Add a hotplug.d/iface script that reloads simet-ma when the WAN
interfaces are brought up, and change the initscript to enumerate the
wan and wan6 netdevs in a file and pass the "-i" option to inetup.

Changes SIMET_RUNTIME_DIR in SIMETBox to mode 02770 (setgid), to ensure
we will have all files in there owned by the nicbr-simet group (for g+w
files).  We want to be able to update some files in there when reloading
simet-ma, and that might happen as the nicbr-simet user.

Due to the addition of the "-i" parameter to inetup, this change
requires simet-ma version 0.16.10 or later, otherwise inetup will refuse
to start.