[MDC-864] simetbox-openwrt-simet-ma: unclone vlabel on boot
Call simet_create_vlabel on every boot, to give it a chance of uncloning
an unit that just got a backup restore using data from some other unit.

This is not something that can be handled through uci-defaults, which is
just for sysupgrade or first-boot.

Remove the uci-defaults/99-simet-ma-vlabel script, since it is no longer
needed (and in fact, it might not work well if the final /var/run is not
yet in place).

PS: a bug in the logger lines of 99-simet-ma-vlabel was noticed while it
was being removed, if it is reinstated: the priority parameter must be
prefixed by the correct option; and the availability of /var/run/simet
must be ensured, both the partition and the directory.