[MDC-864] simetbox-openwrt-simet-ma: track all MACs for anti-cloning
It is possible for the user to confuse the get_mac_address.sh script
into switching from one of the device MACs to another.  This can happen
when the user changes br-lan in such a way that its MAC changes, for
example.  Fortunately, this happens very rarely in real world usage of
the measurement devices.

We are unlikely to be able to fix the root cause of this before OpenWRT
20 (and likely, by switching all of ar71xx to ath79).

Meanwhile, track all non-virtual MAC addresses in the anti-cloning
metadata, and consider it cloned only when none of them matches the
return of get_mac_address.sh.

Note that this change *requires* simet_get_persistent_macs() implemented
by simet_lib.sh, which landed in simet-ma v0.16.10.