[MDC-906] simetbox-openwrt-luci: handle a disabled SIMET1
When SIMET1 is disabled, we lose a few Lua modules, and we wouldn't want
to show anything about SIMET1 in the UI anyway.

LuCI controller: register SIMET1 nodes only where available, switch to
new-style local requires for the modules that might not be present.

LuCI views: detect the lack of SIMET1 in the SIMET2 results page and do
not mention the SIMET1 page.