[MDC-902][MDC-907] make SIMET1 runtime dependencies conditional
Do not select/depend on libopenssl when SIMET1 is disabled.

Also apply the same logic on libpthreads and simetbox-openwrt-config,
although those have minimal space savings when compared to the OpenSSL
monster.

Note that we still build SIMET1, we just don't install it, so the
now-conditional runtime dependencies must be kept as unconditional
*build* dependencies.