[MDC-902][MDC-906] simetbox-openwrt-luci: unbreak SIMET1 results/config
Some of the changes done to the LuCI pages on how simet_utils and other
simet modules-that-apparently-aren't load broke the whole thing badly.

SIMET1 is on its way out, so we can just revert to the old way of using
the global namespace for now.

Also, remove left-over cruft that was not referenced anywhere else in
the htm pages, on the grounds that they were slowing down things for no
good reason.