[MDC-845] simetbox-openwrt-simet-{ma,lmapd}: lock to specific version
We were using a branch name, sometimes master, sometimes stable, to get
the latest commit of that branch for minor fixes during development.

Unfortunately, this is a fail-unsafe when building production images,
so lock to the specific version.  One has to tag (and push the tag)
simet-ma and simet-lmapd, then update simetbox-openwrt-feed's Makefiles
appropriately, in order to get newer code into simetbox.

The devel branch can do something different, if it wants.

Note that we did *not* do the same treatment for simetbox-openwrt-base
and simetbox-openwrt-config: those still depend on the correct code on
the correct branch, for other reasons that should go away soon enough.

This fixes some version numbering insanity that ended up creeping into
626/626004, where simet-ma 0.16.13 ended up being packaged (for openwrt)
as simetbox-openwrt-simet-ma-0.16.12-1.