[MDC-980] simetbox-openwrt-simet-ma: rpcd: return engine config
Expand ubus rpcd call "simet_ma simet_engine_status" to return a
"local_measurement_config" object with the relevant entries from uci.

We need to do that because "uci" is effectively root-only due to some
sort of either a long-standing bug or design shortcoming that forces the
/etc/config/ file to mode 0600 on "uci commit".

This also changes some whitespace on wan_status output.