[MDC-980] simetbox-openwrt-simet-ma: add UCI config
Ship the base uci config in /etc/config/simet_ma, otherwise we'd need to
create it from scratch using an uci-defaults script.

We will eventually need the uci-defaults script to add non-optional new
config entries to existing systems, etc.  But that's expected, while
creating a config from scratch in uci-defaults isn't that common.

Mark "/etc/config/simet_ma" as a conffile.

NOTE: it would be much better if /etc/config/simet_ma could be mode 0644
so that user nicbr-simet can read it, but "uci commit" always resets
that mode to 0600 and $(INSTALL_CONF) forces mode 0600 as a hint.  It is
not worth it to fight this battle, we will work around it with rpcd.