[MDC-654] simetbox-openwrt-simet-ma: add simet-spoofer support
Add simet_spoofer as an optional, configurable component (in Network ->
SIMETBox), and enable its build on simet-ma when requested.

simet_spoofer requires simet_nping to work.  While it is possible to
compile simet-nping to not require openssl, that disables echo-mode
authentication, thus it is currently not an option for us.  This means
we need to restrict simet-nping (and therefore simet_spoofer) to
SIMETBoxes with openssl, only.

Note that this change requires simet-ma 0.16.13 or later (i.e. with
simet-spoofer support) to actually do anything.