[MDC-980][MDC-848] simetbox-openwrt-simet-ma: add ixbr_ping uci handling
Add an uci session for ixbr_ping and also to the rpcd service.

Note that we purposefully did not add it to the LuCI UI yet: this is
more of an internal measurement right now.  If it gets added to the
public report dashboard, we will consider making its period
configurable.