[MDC-1031] simetbox-openwrt-simet-ma: app_link initial-setup-ubus
Move the (unprivileged) information about whether the SIMETBox is still
in its initial configuration from the static app_link toplevel JSON to a
dynamic UBUS endpoint, which makes it trivial to have it correct in
real-time.

Add the simetbox module, and a public_info method.  Also add a
capability for this to the top-level app_link JSON.

app_link capability: initial-setup-json 0.1.1