lock down externel package revisions and set download hashes
Latest openwrt-19.07, 21.02 and trunk enforces that you MUST have
PKG_HASH or PKG_MIRROR_HASH (depending on download methot), or risk a
broken build.  Basically, it *always* tries to download unless there is
a HASH defined for that download AND it matches what has already been
downloaded.

If there is no PKG_HASH/PKG_MIRROR_HASH defined, it ignores the download
folder and tries to redownload during build (not just during make
download).  If there is no network access, or github is down, etc. it
will fail the build.

Lock down PKG_REV to a specific tag, and add the corresponding PKG_HASH
and PKG_MIRROR_HASH to all packages external to the feed.