[MDC-1154] rpc: simet-ma: add simet_pairing_status call
Add a simet_pairing_status call to ubus rpc simet-ma, which returns a
JSON object:

{ "status": 0,
  "result": {
    "paired":<boolean>,
    "name": <string or null>
    "short_name": <string or null>
  }
}

or an error object:
{ "status": <number>, error: "error description" }

Note that it is possible that the "name" and "short_name" of the
participant that the MA is paired to are unknown or undefined, in which
case they will be either null or missing from the result object.

In case of error, an rpcd status 4 error is returned.

This information is obtained directly from SIMET engine internals (the
pairing information cache file, and the persistent pairing-id file) at
the moment.  This information is updated by simet_register_ma.sh.

In particular, if simet_register_ma.sh has not run yet since boot, the
engine can only inform whether the MA was paired or not: the pairing
attributes (such as name and short_name) are not persisted in FLASH.

A present, but empty pairing information cache file is understood as "MA
is unpaired", and does not cause an error.  An non-empty pairing-id file
is also checked for, and signals that the MA is paired regardless of the
existence of the pairing information cache.

This new ubus rpc simet-ma call requires simet-ma 0.17.7 or later,
otherwise it will always return status 4 on unpaired MAs.