[MDC-1031] app-link: move rpcd simetbox to app-link
The "simetbox" ubus rpc module is used only by app-link, so move it
under applink-files/

Likewise, we only need check_dfl_passwd.lua for the "simetbox" ubus rpc
module, so move it as well.

For now, the ACLs are left in the main module, together with the ones
for simet-ma.

Fixes: commit 106a301335dee553b83c13106db9874391d99ef8
       [MDC-1031] simetbox-openwrt-simet-ma: make applink optional