[MDC-1170] simetbox-openwrt-simet-{ma,lmapd}: use emergency schedule
Use the new emergency schedule (replaces the old default schedule, which
we were still trying to symlink to, *erroneously*), so that there will
always be a schedule for lmapd to start.

Enforce onwership of the schedule file to nicbr-simet:nicbr-simet,
symlink-aware, in the initscripts and cronjobs.  A few places were not
doing it correctly.