[MDC-1329] simetbox-openwrt-simet-ma: rpcd: update comments
In OpenWRT 21+, there are root-only requirements in place on
initscripts, and even if we fix that hackjob, one needs ubus acl.d
access to "ubus call service" which is quite powerful on its own right.

So document what needs root, and for which reason.  Note that we have
to actually unbreak simetbox_lmap-fetch-schedule.sh because it cannot
require root.  This is done elsewhere.