[MDC-1330] firewall-urpf: depend on firewall
firewall-urpf depends on fw3 (package "firewall") to work, so it should
either depend, or select it.

Since OpenWRT 21.02 and earlier uses package "firewall" by default, and
OpenWRT 22.03 and later uses package "firewall4" by default, it is best
to depend on the "firewall" package instead of selecting it.

A new "firewall4-urpf" nftables-based package will be provided for use
with fw4 on OpenWRT 22.03 and later.