[MDC-1330] firewall4-urpf: rework for a default of accept
In some situations (e.g. /etc/init.d/firewall reload), openwrt will
flush all the chains in the "inet fw4" table, but the chains will not be
deleted.

Evidently, this does not have a happy ending when there are hooked
chains with a policy of drop, and they end up empty (thus, dropping all
traffic).

Work around this extremely hazardous pitfall by using a policy of
"accept" for the urpf chain (which is hooked to prerouting).  Explicitly
drop (and count) packets at the end of the urpf chain, now that it will
have a policy of "accept".