[MDC-1405] simetbox-openwrt-simet-{ma,lmapd}: enable service on postinst
OpenWRT 21+ (and perhaps some other versions) will not enable by
default a service that is installed through opkg, unless it is listed
in /etc/services_wanted.

Manually enable the simet-ma and simet-lmapd services in their
respective package postinst scripts.

We do not *start* them, though.  That requires a reboot at the moment,
since at least simet-ma has a "boot" section in the initscript.